Refactor activity events for docs usage
