+++ examples/recipes/backend/demo-server/src/app.ts
@@ -7,6 +7,7 @@ import { startGeneratingMonitoringMetrics } from './monitoring-metrics'
 import { startProcessingBackgroundJobs } from './background-job-service'
 import { startGeneratingChatLogBotMessages } from './chat-room-bot'
 import { batchInsertOrders } from './commerce-orders-generation'
+import { startGeneratingActivityEvents } from './activity-events'
 
 const API_PORT = parseInt(process.env.DEMO_APP_PORT ?? '3123')
 const pgPool = new Pool({
@@ -26,6 +27,7 @@ async function main(): Promise<void> {
   void startGeneratingMonitoringMetrics(pgPool)
   void startProcessingBackgroundJobs(pgPool)
   void startGeneratingChatLogBotMessages(pgPool)
+  void startGeneratingActivityEvents(pgPool)
   void batchInsertOrders(pgPool)
 }
