+++ examples/recipes/src/activity_events/ActivityEventsExample.tsx
@@ -1,9 +1,8 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import { Box, Button, Container } from '@mui/material'
+import { Box, Typography, Container } from '@mui/material'
 import { NavigationBar } from '../components/NavigationBar'
 import { useEffect, useState } from 'react'
 import { useElectric } from '../electric/ElectricWrapper'
-import { generateActivity } from './utilities'
 import { ActivityToast } from './ActivityToast'
 import { ActivityPopover } from './ActivityPopover'
 import { LoadingView } from '../components/LoadingView'
@@ -24,12 +23,6 @@ export const ActivityEventsExample = () => {
     syncItems()
   }, [])
 
-  const generateUserActivity = () => {
-    db.activity_events.create({
-      data: generateActivity(),
-    })
-  }
-
   return (
     <Box sx={{ display: 'flex', flexDirection: 'column', height: '100%' }}>
       <NavigationBar title="Activity Events" items={[<ActivityPopover key="notifications" />]} />
@@ -42,9 +35,7 @@ export const ActivityEventsExample = () => {
             justifyContent: 'center',
             height: '100%',
           }}>
-          <Button variant="contained" size="large" onClick={generateUserActivity}>
-            Generate activity
-          </Button>
+          <Typography variant="h5">Activities will be streaming in</Typography>
           <ActivityToast />
         </Container>
       </LoadingView>
